--- src/main/java/blackjack/domain/CardSymbol.java
@@ -1,18 +1,0 @@
-package blackjack.domain;
-
-public enum CardSymbol {
-    SPADE("스페이드"),
-    DIAMOND("다이아몬드"),
-    HEART("하트"),
-    CLUB("클로버");
-
-    private final String value;
-
-    CardSymbol(String value) {
-        this.value = value;
-    }
-
-    public String get() {
-        return value;
-    }
-}
