+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,5 +1,7 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
+
 public class Dealer extends Gamer {
     private static final String name = "딜러";
     private static final int HIT_UPPER_BOUND = 16;
