+++ src/main/java/blackjack/domain/card/CardPicker.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.ArrayList;
 import java.util.Arrays;
