+++ src/main/java/blackjack/domain/card/CardSymbol.java
@@ -1,0 +1,8 @@
+package blackjack.domain.card;
+
+public enum CardSymbol {
+    SPADE,
+    DIAMOND,
+    HEART,
+    CLUB;
+}
