+++ src/main/java/blackjack/domain/card/Cards.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.ArrayList;
 import java.util.HashSet;
