+++ src/main/java/blackjack/view/OutputView.java
@@ -1,7 +1,7 @@
 package blackjack.view;
 
-import blackjack.domain.Card;
 import blackjack.domain.GameResult;
+import blackjack.domain.card.Card;
 
 import java.util.List;
 import java.util.Map;
